requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile --output-file=requirements.txt
#
cachetools==5.5.0
    # via google-auth
certifi==2024.8.30
    # via requests
charset-normalizer==3.4.0
    # via requests
cloudpathlib[gs]==0.20.0
    # via socket-listener (pyproject.toml)
deprecated==1.2.18
    # via
    #   opentelemetry-api
    #   opentelemetry-semantic-conventions
google-api-core[grpc]==2.24.0
    # via
    #   google-cloud-core
    #   google-cloud-pubsub
    #   google-cloud-storage
google-auth==2.37.0
    # via
    #   google-api-core
    #   google-cloud-core
    #   google-cloud-pubsub
    #   google-cloud-storage
google-cloud-core==2.4.1
    # via google-cloud-storage
google-cloud-pubsub==2.28.0
    # via socket-listener (pyproject.toml)
google-cloud-storage==2.19.0
    # via cloudpathlib
google-crc32c==1.6.0
    # via
    #   google-cloud-storage
    #   google-resumable-media
google-resumable-media==2.7.2
    # via google-cloud-storage
googleapis-common-protos[grpc]==1.66.0
    # via
    #   google-api-core
    #   grpc-google-iam-v1
    #   grpcio-status
grpc-google-iam-v1==0.14.1
    # via google-cloud-pubsub
grpcio==1.71.0
    # via
    #   google-api-core
    #   google-cloud-pubsub
    #   googleapis-common-protos
    #   grpc-google-iam-v1
    #   grpcio-status
grpcio-status==1.71.0
    # via
    #   google-api-core
    #   google-cloud-pubsub
idna==3.10
    # via requests
importlib-metadata==8.5.0
    # via opentelemetry-api
markdown-it-py==3.0.0
    # via rich
mdurl==0.1.2
    # via markdown-it-py
opentelemetry-api==1.30.0
    # via
    #   google-cloud-pubsub
    #   opentelemetry-sdk
    #   opentelemetry-semantic-conventions
opentelemetry-sdk==1.30.0
    # via google-cloud-pubsub
opentelemetry-semantic-conventions==0.51b0
    # via opentelemetry-sdk
proto-plus==1.25.0
    # via
    #   google-api-core
    #   google-cloud-pubsub
protobuf==5.29.1
    # via
    #   google-api-core
    #   google-cloud-pubsub
    #   googleapis-common-protos
    #   grpc-google-iam-v1
    #   grpcio-status
    #   proto-plus
pyasn1==0.6.1
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.4.1
    # via google-auth
pygments==2.18.0
    # via rich
pyyaml==6.0.2
    # via socket-listener (pyproject.toml)
requests==2.32.3
    # via
    #   google-api-core
    #   google-cloud-storage
rich==13.9.4
    # via
    #   rich-argparse
    #   socket-listener (pyproject.toml)
rich-argparse==1.6.0
    # via socket-listener (pyproject.toml)
rsa==4.9
    # via google-auth
typing-extensions==4.12.2
    # via opentelemetry-sdk
urllib3==2.2.3
    # via requests
wrapt==1.17.2
    # via deprecated
zipp==3.21.0
    # via importlib-metadata
